refactor(fe): 멘션 위치 잡는 함수 util로 옮김 #236
